Insertar funcion al main js
