+++ src/index.html
@@ -8,9 +8,15 @@
     <link rel="stylesheet" href="style.css" />
   </head>
   <body>
+    <div id="header" >
     <header>
         <img src="logo.png" alt="">
     </header>
+  </div>
+    <main>
+        <div id = "poke">
+          </div>
+    </main>
       <script type="module" src="main.js"></script>
   </body>
 </html>
